requirements.txt
streamlit
google-generativeai
python-docx
pymupdf
python-dotenv
